scripts/requirements.txt
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
